remove inline tests
